Standardize confirmation function in setup scripts

+++ common_install.sh
@@ -6,59 +6,64 @@
 
 source ~/.dotfiles/define_colours.sh
 
+# Function to ask for confirmation with a description
+function ask_confirmation() {
+    local description="$1"
+    read -p "Do you want to run the following section? (y/n): $description : " choice
+    case "$choice" in
+        [Yy]* ) return 0;;
+        * ) return 1;;
+    esac
+}
+
 # Ask for the administrator password upfront
 sudo -v
 
 # ======================================
-# Symllink to various dotfiles
+# Symllink to various dotfiles and directories
 # ======================================
-ln -s ~/.dotfiles/.gitconfig ~/.gitconfig
-ln -s ~/.dotfiles/.gitignore_global ~/.gitignore_global
+if ask_confirmation "Symmlink various dotfiles"; then
+    ln -s ~/.dotfiles/.gitconfig ~/.gitconfig
+    ln -s ~/.dotfiles/.gitignore_global ~/.gitignore_global
 
-ln -sf ~/.dotfiles/.zshrc ~/.zshrc
+    ln -sf ~/.dotfiles/.zshrc ~/.zshrc
+    ln -sf ~/.dotfiles/.pandoc ~/.pandoc
 
-# =====
-# Move directories to ~/
-# =====
-mv ~/.dotfiles/.pandoc ~/.pandoc
-
-# ==============================
-# Run OS specific install script	
-# ==============================
-unameOut="$(uname -s)"
-case "${unameOut}" in
-    Linux*)     machine=Linux;;
-    Darwin*)    machine=Mac;;
-    *)          machine="UNKNOWN:${unameOut}"
-esac
-echo -e "Running on ${machine}"
-
-# Check the current operating system and source the appropriate script
-if [[ "${machine}" == "Mac" ]]; then
-    source mac_install.sh
-elif [[ "${machine}" == "Linux" ]]; then
-    source linux_install.sh
-else
-    echo "Operating system is not macOS or Linux. Skipping installation."
 fi
 
+# ==============================
+# Run OS specific install script    
+# ==============================
+if ask_confirmation "Run OS specific install script"; then
+    unameOut="$(uname -s)"
+    case "${unameOut}" in
+        Linux*)     machine=Linux;;
+        Darwin*)    machine=Mac;;
+        *)          machine="UNKNOWN:${unameOut}"
+    esac
+    echo -e "Running on ${machine}"
+
+    # Check the current operating system and source the appropriate script
+    if [[ "${machine}" == "Mac" ]]; then
+        source mac_install.sh
+    elif [[ "${machine}" == "Linux" ]]; then
+        source linux_install.sh
+    else
+        echo "Operating system is not macOS or Linux. Skipping installation."
+    fi
+fi
 
 # ==================================
 # Setup zsh and nice-to-have plugins
 # MacOS already has zsh installed
 # ==================================
-# Change zsh to be the default shell
-echo -ne "$BOLD Set zsh to default shell? (y/n): $RESET"
-read -p " " set_zsh
-if [[ $set_zsh == [yY] ]]; then
+if ask_confirmation "Set zsh to default shell"; then
     echo -e "$RESET[D] Sudo-ing to make zsh default shell$RESET$RED"
     sudo sh -c "echo $(which zsh) >> /etc/shells" && chsh -s $(which zsh)
     echo -e "$RESET"
 fi
 
-echo -ne "$BOLD Install zsh plugins? (y/n): $RESET"
-read -p " " set_zsh
-if [[ $set_zsh == [yY] ]]; then
+if ask_confirmation "Install zsh plugins"; then
     echo -e "$RESET[D] Installing zsh plugins$RESET$RED"
     git clone https://github.com/zsh-users/zsh-autosuggestions ~/.zsh/zsh-autosuggestions
     echo -e "$RESET"
